latest updates

+++ Sum1.java
@@ -3,7 +3,7 @@ public class Sum1 {
     public static void main(String[] args) {
         int a = Integer.parseInt(args[0]);
         long b = Long.parseLong(args[1]);
-        float sum = a + b;
+        long sum = a + b;
 
         System.out.println(a + " + " + b + " = " + sum);
     }
